Moving some code to Extension methods
