+++ UnityEngineAnalyzer/UnityEngineAnalyzer/UnityEngineAnalyzer.csproj
@@ -57,7 +57,7 @@
       <DependentUpon>UseCompareTagResources.resx</DependentUpon>
     </Compile>
     <Compile Include="Coroutines\DoNotUseCoroutinesAnalyzer.cs" />
-    <Compile  Include="Coroutines\DoNotUseCoroutinesResources.Designer.cs">
+    <Compile Include="Coroutines\DoNotUseCoroutinesResources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
       <DependentUpon>DoNotUseCoroutinesResources.resx</DependentUpon>
@@ -91,6 +91,7 @@
       <DependentUpon>DoNotUseOnGUIResources.resx</DependentUpon>
     </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RolsynExtensions.cs" />
     <Compile Include="StringMethods\DoNotUseStringMethodsAnalyzer.cs" />
     <Compile Include="StringMethods\DoNotUseStringMethodsResources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -138,7 +139,7 @@
     <EmbeddedResource Include="StringMethods\DoNotUseStringMethodsResources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>DoNotUseForEachInUpdateResources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>    
+    </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
     <None Include="Diagnostic.nuspec">
